update rollapp api
