+++ x/rollapp/client/cli/query.go
@@ -31,8 +31,6 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(CmdShowStateInfo())
 	cmd.AddCommand(CmdListLatestStateInfoIndex())
 	cmd.AddCommand(CmdShowLatestStateInfoIndex())
-	cmd.AddCommand(CmdListBlockHeightToFinalizationQueue())
-	cmd.AddCommand(CmdShowBlockHeightToFinalizationQueue())
 	cmd.AddCommand(CmdShowLatestFinalizedStateInfo())
 
 	// this line is used by starport scaffolding # 1
