+++ x/rollapp/keeper/grpc_query_state_info.go
@@ -2,6 +2,8 @@ package keeper
 
 import (
 	"context"
+	"fmt"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -16,7 +18,7 @@ func (k Keeper) StateInfoAll(c context.Context, req *types.QueryAllStateInfoRequ
 		return nil, status.Error(codes.InvalidArgument, "invalid request")
 	}
 
-	var stateInfos []types.StateInfo
+	var stateInfos []types.StateInfoSummary
 	ctx := sdk.UnwrapSDKContext(c)
 
 	store := ctx.KVStore(k.storeKey)
@@ -27,8 +29,14 @@ func (k Keeper) StateInfoAll(c context.Context, req *types.QueryAllStateInfoRequ
 		if err := k.cdc.Unmarshal(value, &stateInfo); err != nil {
 			return err
 		}
-
-		stateInfos = append(stateInfos, stateInfo)
+		if stateInfo.StateInfoIndex.RollappId == req.RollappId {
+			stateInfoSummary := types.StateInfoSummary{
+				StateInfoIndex: stateInfo.StateInfoIndex,
+				Status:         stateInfo.Status,
+				CreationHeight: stateInfo.CreationHeight,
+			}
+			stateInfos = append(stateInfos, stateInfoSummary)
+		}
 		return nil
 	})
 
@@ -43,16 +51,156 @@ func (k Keeper) StateInfo(c context.Context, req *types.QueryGetStateInfoRequest
 	if req == nil {
 		return nil, status.Error(codes.InvalidArgument, "invalid request")
 	}
+	if req.Height != 0 && req.Index != 0 {
+		return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("only one flag can be use for %d or %d", req.Index, req.Height))
+	}
 	ctx := sdk.UnwrapSDKContext(c)
 
-	val, found := k.GetStateInfo(
-		ctx,
-		req.RollappId,
-		req.Index,
-	)
-	if !found {
-		return nil, status.Error(codes.NotFound, "not found")
+	var stateInfo types.StateInfo
+	if req.Index != 0 {
+		val, found := k.GetStateInfo(
+			ctx,
+			req.RollappId,
+			req.Index,
+		)
+		if !found {
+			return nil, status.Error(codes.NotFound, "not found")
+		}
+		stateInfo = val
+	} else if req.Height != 0 {
+		val, err := k.FindStateInfoByHeight(ctx, req.RollappId, req.Height)
+		if err != nil {
+			return nil, err
+		}
+		stateInfo = *val
 	}
 
-	return &types.QueryGetStateInfoResponse{StateInfo: val}, nil
+	return &types.QueryGetStateInfoResponse{StateInfo: stateInfo}, nil
+}
+
+func (k Keeper) FindStateInfoByHeight(ctx sdk.Context, rollappId string, height uint64) (*types.StateInfo, error) {
+	// check that height not zero
+	if height == 0 {
+		return nil, types.ErrInvalidHeight
+	}
+
+	_, found := k.GetRollapp(ctx, rollappId)
+	if !found {
+		return nil, types.ErrUnknownRollappID
+	}
+
+	stateInfoIndex, found := k.GetLatestStateInfoIndex(ctx, rollappId)
+	if !found {
+		return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+			"LatestStateInfoIndex wasn't found for rollappId=%s",
+			rollappId)
+	}
+	// initial interval to search in
+	startInfoIndex := uint64(1) // see TODO bellow
+	endInfoIndex := stateInfoIndex.Index
+
+	// get state info
+	LatestStateInfo, found := k.GetStateInfo(ctx, rollappId, endInfoIndex)
+	if !found {
+		return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+			"StateInfo wasn't found for rollappId=%s, index=%d",
+			rollappId, endInfoIndex)
+	}
+
+	// check that height exists
+	if height >= LatestStateInfo.StartHeight+LatestStateInfo.NumBlocks {
+		return nil, sdkerrors.Wrapf(types.ErrStateNotExists,
+			"rollappId=%s, height=%d",
+			rollappId, height)
+	}
+
+	// check if the height belongs to this batch
+	if height >= LatestStateInfo.StartHeight {
+		return &LatestStateInfo, nil
+	}
+
+	maxNumberOfSteps := endInfoIndex - startInfoIndex + 1
+	stepNum := uint64(0)
+	for ; stepNum < maxNumberOfSteps; stepNum += 1 {
+		// we know that endInfoIndex > startInfoIndex
+		// otherwise the height should have been found
+		if endInfoIndex <= startInfoIndex {
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+				"endInfoIndex should be != than startInfoIndex rollappId=%s, startInfoIndex=%d, endInfoIndex=%d",
+				rollappId, startInfoIndex, endInfoIndex)
+		}
+		// 1. get state info
+		startStateInfo, found := k.GetStateInfo(ctx, rollappId, startInfoIndex)
+		if !found {
+			// TODO:
+			// if stateInfo is missing it won't be logic error if history deletion be implemented
+			// for that we will have to check the oldest we have
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+				"StateInfo wasn't found for rollappId=%s, index=%d",
+				rollappId, startInfoIndex)
+		}
+		endStateInfo, found := k.GetStateInfo(ctx, rollappId, endInfoIndex)
+		if !found {
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+				"StateInfo wasn't found for rollappId=%s, index=%d",
+				rollappId, endInfoIndex)
+		}
+		startHeight := startStateInfo.StartHeight
+		endHeight := endStateInfo.StartHeight + endStateInfo.NumBlocks - 1
+
+		// 2. check startStateInfo
+		if height >= startStateInfo.StartHeight &&
+			(startStateInfo.StartHeight+startStateInfo.NumBlocks) > height {
+			return &startStateInfo, nil
+		}
+
+		// 3. check endStateInfo
+		if height >= endStateInfo.StartHeight &&
+			(endStateInfo.StartHeight+endStateInfo.NumBlocks) > height {
+			return &endStateInfo, nil
+		}
+
+		// 4. calculate the average blocks per batch
+		avgBlocksPerBatch := (endHeight - startHeight + 1) / (endInfoIndex - startInfoIndex + 1)
+		if avgBlocksPerBatch == 0 {
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+				"avgBlocksPerBatch is zero!!! rollappId=%s, endHeight=%d, startHeight=%d, endInfoIndex=%d, startInfoIndex=%d",
+				rollappId, endHeight, startHeight, endInfoIndex, startInfoIndex)
+		}
+
+		// 5. load the candidate block batch
+		infoIndexStep := (height - startHeight) / avgBlocksPerBatch
+		if infoIndexStep == 0 {
+			infoIndexStep = 1
+		}
+		candidateInfoIndex := startInfoIndex + infoIndexStep
+		if candidateInfoIndex > endInfoIndex {
+			// skip to the last, probably the steps to big
+			candidateInfoIndex = endInfoIndex
+		}
+		if candidateInfoIndex == endInfoIndex {
+			candidateInfoIndex = endInfoIndex - 1
+		}
+		candidateStateInfo, found := k.GetStateInfo(ctx, rollappId, candidateInfoIndex)
+		if !found {
+			return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+				"StateInfo wasn't found for rollappId=%s, index=%d",
+				rollappId, candidateInfoIndex)
+		}
+
+		// 6. check the candidate
+		if candidateStateInfo.StartHeight > height {
+			endInfoIndex = candidateInfoIndex - 1
+		} else {
+			if candidateStateInfo.StartHeight+candidateStateInfo.NumBlocks-1 < height {
+				startInfoIndex = candidateInfoIndex + 1
+			} else {
+				return &candidateStateInfo, nil
+			}
+		}
+	}
+
+	return nil, sdkerrors.Wrapf(sdkerrors.ErrLogic,
+		"More searching steps than indexes! rollappId=%s, stepNum=%d, maxNumberOfSteps=%d",
+		rollappId, stepNum, maxNumberOfSteps)
 }
